+++ src/main/java/org/sbrubbles/conditio/restarts/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides some general use restarts.
+ */
+package org.sbrubbles.conditio.restarts;