+++ src/test/java/org/sbrubbles/conditio/BasicOperationsTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.sbrubbles.conditio.fixtures.BasicCondition;
-import org.sbrubbles.conditio.fixtures.BasicSignal;
+import org.sbrubbles.conditio.fixtures.BasicNotice;
 import org.sbrubbles.conditio.fixtures.PleaseSignalSomethingElse;
 import org.sbrubbles.conditio.fixtures.SonOfBasicCondition;
 import org.sbrubbles.conditio.fixtures.logging.MalformedLogEntry;
@@ -27,7 +27,7 @@ public class BasicOperationsTest {
     final String TEST_STR = "test";
     final Restart.Option u = new UseValue(TEST_STR);
 
-    try (Scope a = Scope.create()) {
+    try (Scope a = Scopes.create()) {
       // no restart before the handler...
       assertFalse(toStream(a.getAllRestarts()).anyMatch(r -> r.test(u)), "before handle");
 
@@ -41,7 +41,7 @@ public class BasicOperationsTest {
       // no restart after either
       assertFalse(toStream(a.getAllRestarts()).anyMatch(r -> r.test(u)), "after handle");
 
-      try (Scope b = Scope.create()) {
+      try (Scope b = Scopes.create()) {
         // no restart before signal...
         assertFalse(toStream(b.getAllRestarts()).anyMatch(r -> r.test(u)), "before signal");
 
@@ -59,7 +59,7 @@ public class BasicOperationsTest {
     final String TEST_STR = "test";
     final Restart.Option u = new UseValue(TEST_STR);
 
-    try (Scope a = Scope.create()) {
+    try (Scope a = Scopes.create()) {
       assertFalse(toStream(a.getAllRestarts()).anyMatch(r -> r.test(u)), "before handle");
 
       a.handle(MalformedLogEntry.class, (c, ops) -> {
@@ -70,7 +70,7 @@ public class BasicOperationsTest {
 
       assertEquals(TEST_STR, a.call(
         () -> {
-          try (Scope b = Scope.create()) {
+          try (Scope b = Scopes.create()) {
             assertTrue(toStream(b.getAllRestarts()).anyMatch(r -> r.test(u)), "inside call");
 
             return b.signal(new MalformedLogEntry(""));
@@ -88,19 +88,19 @@ public class BasicOperationsTest {
     final String EXPECTED_RESULT = "<result>";
     final List<String> trace = new ArrayList<>();
 
-    try (Scope a = Scope.create()) {
+    try (Scope a = Scopes.create()) {
       a.handle(BasicCondition.class, (c, ops) -> {
         trace.add("a");
         return ops.use(EXPECTED_RESULT);
       });
 
-      try (Scope b = Scope.create()) {
+      try (Scope b = Scopes.create()) {
         b.handle(BasicCondition.class, (c, ops) -> {
           trace.add("b");
           return ops.skip();
         });
 
-        try (Scope c = Scope.create()) {
+        try (Scope c = Scopes.create()) {
           Object actual = c.signal(builder.apply(EXPECTED_RESULT));
 
           assertEquals(EXPECTED_RESULT, actual);
@@ -117,14 +117,14 @@ public class BasicOperationsTest {
     final String EXPECTED_RESULT = "<result>";
     final List<String> trace = new ArrayList<>();
 
-    try (Scope a = Scope.create()) {
+    try (Scope a = Scopes.create()) {
       a.handle(BasicCondition.class,
         (c, ops) -> {
           trace.add("a");
           return ops.use(EXPECTED_RESULT);
         });
 
-      try (Scope b = Scope.create()) {
+      try (Scope b = Scopes.create()) {
         Object actual = b.signal(new BasicCondition(""));
 
         assertEquals(EXPECTED_RESULT, actual);
@@ -138,23 +138,23 @@ public class BasicOperationsTest {
     final String FIXED_RESULT = "<result>";
     final List<String> trace = new ArrayList<>();
 
-    try (Scope a = Scope.create()) {
+    try (Scope a = Scopes.create()) {
       a.handle(PleaseSignalSomethingElse.class,
         (c, ops) -> {
           trace.add("a");
-          try (Scope s = Scope.create()) {
+          try (Scope s = Scopes.create()) {
             return ops.use(s.signal(new BasicCondition(null)));
           }
         });
 
-      try (Scope b = Scope.create()) {
+      try (Scope b = Scopes.create()) {
         b.handle(BasicCondition.class,
           (c, ops) -> {
             trace.add("b");
             return ops.use(FIXED_RESULT);
           });
 
-        try (Scope c = Scope.create()) {
+        try (Scope c = Scopes.create()) {
           Object actual = c.signal(new PleaseSignalSomethingElse());
 
           assertEquals(FIXED_RESULT, actual);
@@ -170,7 +170,7 @@ public class BasicOperationsTest {
   public void signallingAConditionWithNoHandlersErrorsOut() {
     BasicCondition condition = new BasicCondition("test");
 
-    try (Scope a = Scope.create()) {
+    try (Scope a = Scopes.create()) {
       a.signal(condition);
 
       fail();
@@ -180,10 +180,10 @@ public class BasicOperationsTest {
   }
 
   @Test
-  public void signallingASignalWithNoHandlersJustNopesOut() {
-    BasicSignal condition = new BasicSignal("test");
+  public void signallingANoticeWithNoHandlersJustNopesOut() {
+    BasicNotice condition = new BasicNotice("test");
 
-    try (Scope a = Scope.create()) {
+    try (Scope a = Scopes.create()) {
       a.signal(condition);
     }