+++ src/test/java/org/sbrubbles/conditio/ScopeTest.java
@@ -4,15 +4,16 @@ import org.junit.jupiter.api.Test;
 import org.sbrubbles.conditio.fixtures.BasicCondition;
 import org.sbrubbles.conditio.fixtures.logging.UseValue;
 
+import java.util.Iterator;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 public class ScopeTest {
   @Test
   public void rootHasNullParent() {
-    try(Scope scope = Scope.create()) {
-      assertTrue(scope.isRoot());
+    try (Scope scope = Scopes.create()) {
       assertNull(scope.getParent());
     }
   }
@@ -21,29 +22,27 @@ public class ScopeTest {
   public void everyInvocationChainHasItsOwnRoot() {
     Scope firstTry;
 
-    try(Scope scope = Scope.create()) {
-      assertTrue(scope.isRoot());
+    try (Scope scope = Scopes.create()) {
+      assertNull(scope.getParent());
 
       firstTry = scope;
     }
 
-    try(Scope scope = Scope.create()) {
-      assertTrue(scope.isRoot());
+    try (Scope scope = Scopes.create()) {
+      assertNull(scope.getParent());
       assertNotEquals(firstTry, scope);
     }
   }
 
   @Test
   public void createReflectsTheTryStack() {
-    try(Scope a = Scope.create()) {
-      assertTrue(a.isRoot());
+    try (Scope a = Scopes.create()) {
+      assertNull(a.getParent());
 
-      try(Scope b = Scope.create()) {
-        assertFalse(b.isRoot());
+      try (Scope b = Scopes.create()) {
         assertEquals(b.getParent(), a);
 
-        try (Scope c = Scope.create()) {
-          assertFalse(c.isRoot());
+        try (Scope c = Scopes.create()) {
           assertEquals(c.getParent(), b);
         }
       }
@@ -52,7 +51,7 @@ public class ScopeTest {
 
   @Test
   public void getHandlersIsUnmodifiable() {
-    try (Scope a = Scope.create()) {
+    try (Scope a = Scopes.create()) {
       final List<Handler> hs = a.getHandlers();
       final Handler h = new HandlerImpl(BasicCondition.class, (c, ops) -> ops.use("test"));
 
@@ -63,12 +62,31 @@ public class ScopeTest {
 
   @Test
   public void getRestartsIsUnmodifiable() {
-    try (Scope a = Scope.create()) {
+    try (Scope a = Scopes.create()) {
       final List<Restart> rs = a.getRestarts();
       final Restart r = new RestartImpl(UseValue.class, u -> "test");
 
       assertThrows(UnsupportedOperationException.class, () -> rs.add(r));
       assertThrows(UnsupportedOperationException.class, () -> rs.remove(r));
+    }
+  }
+
+  @Test
+  public void theIteratorFollowsTheProperProtocol() {
+    try (Scope a = Scopes.create()) {
+      Iterator<Handler> iterator = a.getAllHandlers().iterator();
+
+      assertFalse(iterator.hasNext());
+      assertThrows(NoSuchElementException.class, () -> iterator.next());
+    }
+  }
+
+  @Test
+  public void signalThrowsIfHandlerReturnsNull() {
+    try (Scope a = Scopes.create()) {
+      a.handle(BasicCondition.class, (c, ops) -> null);
+
+      assertThrows(UnsupportedOperationException.class, () -> a.signal(new BasicCondition("oops")));
     }
   }
 }