+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 * Handlers now take a Handler.Operations instead of a scope, which delimits the operations available.
 * Handlers now return a Handler.Decision, which is written in a way to effectively force the dev to use Handler.Operations to get an instance.
-* Condition now provides a protocol for Scope.signal. It can be overridden, paving the way for Signal, Warning and other possible Condition subtypes.
-* Reorganizing some tests to improve legibility.
-* Adding some possible usages in the tests. 
+* Scope is now an interface, with the Scopes class managing the stack.
+* Condition is now a class which provides a callback for Scope.signal, paving the way for Notice (the RuntimeException to Condition's Exception), Warning and other possible subtype protocols.
+* Provided some general use restart options, like Resume.
+* Reorganizing the tests to improve legibility, and adding some possible usages.
 
 # 0.3